requirements.txt
openai_whisper==20230314
PyAudio==0.2.13
pyttsx3==2.90
Requests==2.31.0
SpeechRecognition==3.10.0
torch==2.0.1
